requirements.txt
google-cloud-aiplatform
google-generativeai
ipython
rel
requests
websocket-client
